changed /dashboard/group/[groupId] to tsx
